+++ internal/common/validation/validations.go
@@ -1,4 +1,4 @@
-package validations
+package validation
 
 import (
 	"fmt"
